refactor(obd2): split connect classifier + move trace card + lint fixes (#2969)
Post-merge lint-gate fixes for the #2969 connect-tracing work:

- Split the two free classifier functions (classifyObd2ConnectError /
  classifyBleOpenOutcome) into obd2_connect_classifier.dart so
  obd2_connect_trace_log.dart drops back under the 400-line cap.
- Move Obd2ConnectTraceCard into presentation/widgets/ (a reusable card
  primitive, like Obd2DiagnosticsCard) so it's out of the SectionCard-only
  presentation/screens/ scope (#923 no-raw-Card lint).
- Factor the channel-open trace stamp into the shared
  Obd2ConnectTraceLog.stampOpenFailure one-liner (BLE + Classic channels).
- catch_no_st (#1103): the connect-trace `catch (e)` arms are rethrow-only with
  a single-line `// ignore: catch_no_st`; the recoverable by-MAC catches bind
  `(e, st)` + debugPrint under assert; endTrace's never-throws catches bind
  `(e, st)` + debugPrint (no more silent `catch (_) {}`).
- #2349 never-throws: added the endTrace fault-injection test (a throwing
  onTraceAdded listener → endTrace returns normally + the trace still lands).
- file_length grandfathers (decomposition tracked by #2190): adapter_registry
  500→522 (transportForName), obd2_connection_service →563 (per-method trace
  wrappers), flutter_blue_plus_elm_channel →419 (scan-path timeout +
  channel-open stamp).
