+++ test/features/profile/presentation/screens/developer_tools/obd2_health_connect_traces_test.dart
@@ -9,7 +9,7 @@ import 'package:tankstellen/features/consumption/data/obd2/obd2_connect_trace.da
 import 'package:tankstellen/features/consumption/data/obd2/obd2_connect_trace_log.dart';
 import 'package:tankstellen/features/feature_management/application/feature_flags_provider.dart';
 import 'package:tankstellen/features/feature_management/domain/feature.dart';
-import 'package:tankstellen/features/profile/presentation/screens/developer_tools/obd2_connect_trace_card.dart';
+import 'package:tankstellen/features/consumption/presentation/widgets/obd2_connect_trace_card.dart';
 import 'package:tankstellen/features/profile/presentation/screens/developer_tools/obd2_health_screen.dart';
 import 'package:tankstellen/features/vehicle/domain/entities/vehicle_profile.dart';
 import 'package:tankstellen/features/vehicle/providers/vehicle_providers.dart';
